Refactor: Modularize APIService.ts
Separated concerns from the large APIService.ts file into multiple, more focused services:

- ValidationService: Handles validation of AI findings.
- ConfidenceScorer: Manages scoring of AI finding confidence.
- DataFetchingService: Responsible for fetching data from external sources like NVD and EPSS.
- AIEnhancementService: Contains methods for AI-powered threat intelligence, patch/advisory search, and analysis generation.
- UtilityService: Houses common helper functions used across services.

APIService.ts now acts as a facade, orchestrating calls to these new specialized services. This improves maintainability and readability of the codebase.
